Artifacts from https://github.com/floodlight/loxigen/compare/b082bcd13a0bb8270df593d3ae468b95450a55a5...7b7c3619cf68d3654b301cc92efd4341cddfe323
Loxigen Head commit floodlight/loxigen@7b7c3619cf68d3654b301cc92efd4341cddfe323
commit 7b7c3619cf68d3654b301cc92efd4341cddfe323
Merge: b082bcd 2a66845
Author: abat <abat@bigswitch.com>
Date:   Mon Jun 9 17:21:37 2014 -0700

    Merge into master from pull request #280:
    bsn_flow_checksum: change table_id type to 8 bits for consistency (https://github.com/floodlight/loxigen/pull/280)

commit 2a668453807816185a9b303fc7ec109cd8a396b1
Author: Rich Lane <rlane@bigswitch.com>
Date:   Mon Jun 9 17:14:26 2014 -0700

    bsn_flow_checksum: change table_id type to 8 bits for consistency

    Compatibility is maintained for valeus that fit in 8 bits.

+++ loci/inc/loci/loci_classes.h
@@ -13638,10 +13638,10 @@ extern void of_bsn_table_set_buckets_size_subtype_get(
 
 extern void of_bsn_table_set_buckets_size_table_id_set(
     of_bsn_table_set_buckets_size_t *obj,
-    uint16_t table_id);
+    uint8_t table_id);
 extern void of_bsn_table_set_buckets_size_table_id_get(
     of_bsn_table_set_buckets_size_t *obj,
-    uint16_t *table_id);
+    uint8_t *table_id);
 
 extern void of_bsn_table_set_buckets_size_buckets_size_set(
     of_bsn_table_set_buckets_size_t *obj,
